+++ api-veterinary/app/Http/Controllers/Staff/StaffController.php
@@ -2,14 +2,15 @@
 
 namespace App\Http\Controllers\Staff;
 
-use App\Http\Controllers\Controller;
-use App\Http\Resources\User\UserCollection;
-use App\Http\Resources\User\UserResource;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Storage;
 use Spatie\Permission\Models\Role;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Storage;
+use App\Http\Resources\User\UserResource;
+use App\Http\Resources\User\UserCollection;
 
 class StaffController extends Controller
 {
@@ -18,15 +19,18 @@ class StaffController extends Controller
      */
     public function index(Request $request)
     {
+        Gate::authorize("viewAny",User::class);
         $search = $request->get("search");
-        $users = User::where(DB::raw("users.name || ' ' || COALESCE(users.surname,'')|| ' ' || users.email"), "ilike", "%" . $search . "%")
-            ->whereHas("roles", function ($query) use ($search) {
-                $query->where(DB::raw("users.name || ' ' || COALESCE(users.surname,'')|| ' ' || users.email"), "ilike", "%" . $search . "%");
-            })
-            ->orderBy("id", "asc")->get();
+
+        $users = User::where(DB::raw("users.name || ' ' || COALESCE(users.surname,'') || ' ' || users.email"),"ilike","%".$search."%")
+                ->whereHas("roles",function($q) {
+                    $q->where("name","not ilike","%veterinario%");
+                })        
+                ->orderBy("id","desc")->get();
+
         return response()->json([
             "users" => UserCollection::make($users),
-            "roles" => Role::where("name", "not ilike", "%veterinario%")->get()->map(function ($role) {
+            "roles" => Role::where("name","not ilike","%veterinario%")->get()->map(function($role) {
                 return [
                     "id" => $role->id,
                     "name" => $role->name
@@ -40,26 +44,24 @@ class StaffController extends Controller
      */
     public function store(Request $request)
     {
-        $is_user_exists = User::where("email", $request->email)->first();
-        if ($is_user_exists) {
+        Gate::authorize("create",User::class);
+        $is_user_exists = User::where("email",$request->email)->first();
+        if($is_user_exists){
             return response()->json([
                 "message" => 403,
                 "message_text" => "El usuario ya existe"
             ]);
         }
-
-        if ($request->hasFile("imagen")) {
-            $path = Storage::putFile("users", $request->file("imagen"));
+        if($request->hasFile("imagen")){
+            $path = Storage::putFile("users",$request->file("imagen"));
             $request->request->add(["avatar" => $path]);
         }
-
-        if ($request->password) {
+        if($request->password){
             $request->request->add(["password" => bcrypt($request->password)]);
         }
-        // if ($request->birthday) {
-        //     $request->request->add(["birthday" => $request->birthday . "00:00:00"]);
-        // }
-
+        if($request->birthday){
+            $request->request->add(["birthday" => $request->birthday." 00:00:00"]);
+        }
         $user = User::create($request->all());
         $role = Role::findOrFail($request->role_id);
         $user->assignRole($role);
@@ -83,32 +85,36 @@ class StaffController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $is_user_exists = User::where("email", $request->email)->where("id", "<>", $id)->first();
-        if ($is_user_exists) {
+        Gate::authorize("update",User::class);
+        $is_user_exists = User::where("email",$request->email)->where("id","<>",$id)->first();
+        if($is_user_exists){
             return response()->json([
                 "message" => 403,
                 "message_text" => "El usuario ya existe"
             ]);
         }
-
         $user = User::findOrFail($id);
-
-        if ($request->hasFile("imagen")) {
-            if ($user->avatar) {
+        if($request->hasFile("imagen")){
+            if($user->avatar){
                 Storage::delete($user->avatar);
             }
-            $path = Storage::putFile("users", $request->file("imagen"));
+            $path = Storage::putFile("users",$request->file("imagen"));
             $request->request->add(["avatar" => $path]);
         }
-
-        if ($request->password) {
+        if($request->password){
             $request->request->add(["password" => bcrypt($request->password)]);
         }
-
+        if($request->birthday){
+            $request->request->add(["birthday" => $request->birthday." 00:00:00"]);
+        }
         $user->update($request->all());
 
-        if ($request->role_id && $request->role_id != $user->roles->first()?->id) {
-            $user->syncRoles([Role::findOrFail($request->role_id)]);
+        if($request->role_id && $request->role_id != $user->role_id){
+            $role_old = Role::findOrFail($user->role_id);
+            $user->removeRole($role_old);
+
+            $role_new = Role::findOrFail($request->role_id);
+            $user->assignRole($role_new);
         }
 
         return response()->json([
@@ -122,10 +128,15 @@ class StaffController extends Controller
      */
     public function destroy(string $id)
     {
+        Gate::authorize("delete",User::class);
         $user = User::findOrFail($id);
-        if ($user->avatar) {
+        if($user->avatar){
             Storage::delete($user->avatar);
         }
         $user->delete();
+
+        return response()->json([
+            "message" => 200,
+        ]);
     }
 }
